fix:refactor de movimientos
