+++ src/features/movements/utils/format.ts
@@ -1,0 +1,20 @@
+const amountFormatter = new Intl.NumberFormat('en-US', {
+  style: 'currency',
+  currency: 'USD',
+});
+
+export const formatMovementAmount = (value: string) => {
+  const numeric = Number(value);
+  if (Number.isNaN(numeric)) {
+    return value;
+  }
+  return amountFormatter.format(numeric);
+};
+
+export const formatMovementDate = (value: string) => {
+  const parsed = new Date(value);
+  if (Number.isNaN(parsed.valueOf())) {
+    return value;
+  }
+  return parsed.toLocaleDateString();
+};
